requirements.txt
Django==3.2.6
psycopg2-binary==2.8.5
Pillow==8.0
djangorestframework==3.12.4
python-decouple==3.4
djangorestframework_simplejwt==4.7.2
drf-yasg==1.20
django-filter==2.4.0